remove some warnings and turn hazel into static lib

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -4,8 +4,8 @@
 
 #include "imgui.h"
 #include "GLFW/glfw3.h"
-#include "examples/imgui_impl_glfw.cpp"
-#include "examples/imgui_impl_opengl3.cpp"
+#include <examples/imgui_impl_glfw.h>
+#include <examples/imgui_impl_opengl3.h>
 
 #include "Hazel/Application.h"
 
@@ -81,7 +81,7 @@ namespace Hazel {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
